Add support for bulk user exports
This commit adds the ability to create bulk user exports.  With this
update, a new request type of `StatusCodeRequest` was added in order to
handle queries against the `job_error` endpoint which report the error
result of a given job.

Please note:  With this change, consumers of the API will be responsible
for fetching and decompressing the `gzip` file produced from the job
request via the URL provided in the response.