ManagementAPI: Allow customizing HTTP client
This simple change allows an `OkHttpClient.Builer` to be passed into the ManagementAPI’s constructor allowing clients of the ManagmentAPI to customize the HTTP client that is used internally.

#### Use Case
We are using the ManagementAPI to handle fine-grained permissions with multi-tenancy by adding an “API per account”  (per the Auth0 recommendation).  This requires us to interact twith the Management API regularly. The API has a very low rate limit and we were hitting it constantly.  OkHttp’s `Interceptor`s were a perfect solution but the current code offers no ability to access or customize the HTTP client. This commit allows us to pass in a client builder with a rate limiting interceptor and resolve our issues with rate limiting.