#269 : Missing Management API Functionality for Linking User Accounts (link_with)
The Java Auth0 API is missing the ability to link accounts using the "link_with" body parameter option. This is useful for when the client sends the secondary ID token to link with.

Override the linkIdentity method in UsersEntity.java to support the ability to supply a ID token instead of a connection and provider string. See: https://auth0.com/docs/api/management/v2#!/Users/post_identities