Invitation fails when you want to invite with roles
The Management REST API expects roles to be a JSON array but they are
instead constructed as a roles object.
For example the roles field will look like
```{... "roles" : {"roles" : ["role-1", "role-2"]} ...}```

Due to this api fails with the following error:
com.auth0.exception.APIException: Request failed with status code 400:
Payload validation error: 'Expected type array but found type object'
on property roles (List of roles IDs to associated with the user).

This commit fixes that.